+++ CHANGELOG.md
@@ -19,6 +19,15 @@ Security in case of vulnerabilities.
  - New route `/projects/all/by-month/transfers/{month}/{year}`
  - New route `/projects/all/by-month/conversions/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
  - New route `/projects/all/by-month/transfers/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
+ - New route `/projects/team/new`
+ - New route `/projects/team/handed-over`
+ - Your team projects "new" list (`/projects/team/new`)
+ - Your team projects handed over list (`/projects/team/handed-over`)
+ - Add new `ProjectTeam` extension method `TeamIsRegionalCaseworkServices`, to identify RCS users 
+
+### Changes
+- Add an "orderBy" argument to the `ListAllProjectsByFilter` query
+- Allow `ListAllProjectsByFilter` query to handle multiple filters
  
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-04-24.175...main) for everything awaiting release